+++ src/RhinoInside.Revit.GH/Components/Component.cs
@@ -542,9 +542,9 @@ namespace RhinoInside.Revit.GH.Components
     public virtual bool NeedsToBeExpired
     (
       ARDB.Document document,
-      ICollection<ARDB.ElementId> added,
-      ICollection<ARDB.ElementId> deleted,
-      ICollection<ARDB.ElementId> modified
+      ISet<ARDB.ElementId> added,
+      ISet<ARDB.ElementId> deleted,
+      ISet<ARDB.ElementId> modified
     )
     {
       // Only Query-Collector components need to be expired when something is added.
@@ -558,7 +558,7 @@ namespace RhinoInside.Revit.GH.Components
         // Check inputs
         foreach (var param in persistentInputs)
         {
-          if (param.NeedsToBeExpired(document, ElementIdExtension.EmptyCollection, deleted, modified))
+          if (param.NeedsToBeExpired(document, ElementIdExtension.EmptySet, deleted, modified))
             return true;
         }
       }