+++ src/RhinoInside.Revit.GH/Components/ElementCollectorComponent.cs
@@ -17,9 +17,9 @@ namespace RhinoInside.Revit.GH.Components
     public override bool NeedsToBeExpired
     (
       ARDB.Document document,
-      ICollection<ARDB.ElementId> added,
-      ICollection<ARDB.ElementId> deleted,
-      ICollection<ARDB.ElementId> modified
+      ISet<ARDB.ElementId> added,
+      ISet<ARDB.ElementId> deleted,
+      ISet<ARDB.ElementId> modified
     )
     {
       // Check if the change is on a document this component is querying.
@@ -48,9 +48,9 @@ namespace RhinoInside.Revit.GH.Components
     protected virtual bool MayNeedToBeExpired
     (
       ARDB.Document document,
-      ICollection<ARDB.ElementId> added,
-      ICollection<ARDB.ElementId> deleted,
-      ICollection<ARDB.ElementId> modified
+      ISet<ARDB.ElementId> added,
+      ISet<ARDB.ElementId> deleted,
+      ISet<ARDB.ElementId> modified
     )
     {
       var elementFilter = ElementFilter;
@@ -73,7 +73,7 @@ namespace RhinoInside.Revit.GH.Components
         {
           foreach (var param in Params.Output.OfType<Kernel.IGH_ReferenceParam>())
           {
-            if (param.NeedsToBeExpired(document, ElementIdExtension.EmptyCollection, deleted, ElementIdExtension.EmptyCollection))
+            if (param.NeedsToBeExpired(document, ElementIdExtension.EmptySet, deleted, ElementIdExtension.EmptySet))
               return true;
           }
         }