+++ src/RhinoInside.Revit.GH/Components/ObjectStyles/QueryCategories.cs
@@ -18,9 +18,9 @@ namespace RhinoInside.Revit.GH.Components.ObjectStyles
     protected override bool MayNeedToBeExpired
     (
       ARDB.Document document,
-      ICollection<ARDB.ElementId> added,
-      ICollection<ARDB.ElementId> deleted,
-      ICollection<ARDB.ElementId> modified
+      ISet<ARDB.ElementId> added,
+      ISet<ARDB.ElementId> deleted,
+      ISet<ARDB.ElementId> modified
     )
     {
       if (added.Any(x => x.IsCategoryId(document)))
@@ -33,7 +33,7 @@ namespace RhinoInside.Revit.GH.Components.ObjectStyles
       {
         foreach (var param in Params.Output.OfType<Kernel.IGH_ReferenceParam>())
         {
-          if (param.NeedsToBeExpired(document, ElementIdExtension.EmptyCollection, deleted, ElementIdExtension.EmptyCollection))
+          if (param.NeedsToBeExpired(document, ElementIdExtension.EmptySet, deleted, ElementIdExtension.EmptySet))
             return true;
         }
       }