+++ src/RhinoInside.Revit.GH/Components/Workset/QueryWorksets.cs
@@ -18,9 +18,9 @@ namespace RhinoInside.Revit.GH.Components.Worksets
     public override bool NeedsToBeExpired
     (
       ARDB.Document document,
-      ICollection<ARDB.ElementId> added,
-      ICollection<ARDB.ElementId> deleted,
-      ICollection<ARDB.ElementId> modified
+      ISet<ARDB.ElementId> added,
+      ISet<ARDB.ElementId> deleted,
+      ISet<ARDB.ElementId> modified
     ) => false;
 
     #region UI