+++ src/RhinoInside.Revit.GH/Parameters/Input/DocumentPicker.cs
@@ -71,9 +71,9 @@ namespace RhinoInside.Revit.GH.Parameters.Input
     bool Kernel.IGH_ReferenceParam.NeedsToBeExpired
     (
       ARDB.Document doc,
-      ICollection<ARDB.ElementId> added,
-      ICollection<ARDB.ElementId> deleted,
-      ICollection<ARDB.ElementId> modified
+      ISet<ARDB.ElementId> added,
+      ISet<ARDB.ElementId> deleted,
+      ISet<ARDB.ElementId> modified
     )
     {
       // Since casting from `ARDB.BuiltInCategory` to `Types.Category` assumes current document…
@@ -192,9 +192,9 @@ namespace RhinoInside.Revit.GH.Parameters.Input
     bool Kernel.IGH_ReferenceParam.NeedsToBeExpired
     (
       ARDB.Document doc,
-      ICollection<ARDB.ElementId> added,
-      ICollection<ARDB.ElementId> deleted,
-      ICollection<ARDB.ElementId> modified
+      ISet<ARDB.ElementId> added,
+      ISet<ARDB.ElementId> deleted,
+      ISet<ARDB.ElementId> modified
     )
     {
       // If selected items are modified we need to expire dependant components